Start modularization of the tokenizing process
